encoder: validate geometry area using widened arithmetic (#3976)
- Cast kiPicWidth, kiPicHeight, and MAX_MBS_PER_FRAME to i64 before multiplication in ParamValidationExt
- Prevent integer overflow panic in debug mode and wrapping in release mode
- Add test_encoder_initialize_ext_rejects_geometry_overflow integration test mirroring C++ UT

+++ rust/crates/openh264-rs/src/encoder/wels_encoder_ext.rs
@@ -1212,7 +1212,7 @@ pub fn ParamValidationExt(pLogCtx: SLogContext, pCodingParam: &mut SWelsSvcCodin
 
         if kiPicWidth <= 0
             || kiPicHeight <= 0
-            || kiPicWidth * kiPicHeight > (MAX_MBS_PER_FRAME << 8)
+            || (kiPicWidth as i64) * (kiPicHeight as i64) > ((MAX_MBS_PER_FRAME as i64) << 8)
         {
             return ENC_RETURN_UNSUPPORTED_PARA;
         }
